District input to register view (required) and district input widget
