+++ lib/views/register_view.dart
@@ -4,11 +4,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_ecoapp/bloc/profile_bloc.dart';
+import 'package:flutter_ecoapp/models/district.dart';
 import 'package:flutter_ecoapp/models/profile.dart';
 import 'package:flutter_ecoapp/utils/email_util.dart';
 import 'package:flutter_ecoapp/views/register_pass_view.dart';
 import 'package:flutter_ecoapp/views/style/colors.dart';
 import 'package:flutter_ecoapp/views/widgets/bottom_nav_bar.dart';
+import 'package:flutter_ecoapp/views/widgets/district_input.dart';
 import 'package:flutter_ecoapp/views/widgets/normal_button.dart';
 import 'package:flutter_ecoapp/views/widgets/normal_input.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -35,6 +37,10 @@ class _RegisterViewState extends State<RegisterView> {
 
   final birthday = {
     'birthday': DateTime.now()
+  };
+
+  final Map<String, DistrictModel?> selectedDistrict = {
+    'district': null
   };
 
   final _formKey = GlobalKey<FormState>();
@@ -186,24 +192,18 @@ class _RegisterViewState extends State<RegisterView> {
               controller: controllers['date']!,
               validator: (value) => value!.isEmpty? 'Debe ingresar su fecha de nacimiento' : null
             ),
-            /* NormalInput(
-              header: 'Comuna', 
-              hint: 'Ingresa tu comuna', 
-              icon: Icons.location_on,
-              readOnly: false, // TODO: Change to true
-              onTap: (){
-                
-              },
-              controller: controllers['district']!,
-              validator: (value) => value!.isEmpty? 'Debe ingresar su comuna' : null
+            Divider(thickness: 1,),
+            SizedBox(height: 20.0,),
+            DistrictInput(
+              selectedDistrict: (value) => selectedDistrict['district'] = value,
             ),
             NormalInput(
               header: 'Dirección', 
               hint: 'Ingresa tu dirección', 
               icon: Icons.location_on,
               controller: controllers['location']!,
-              validator: (value) => value!.isEmpty? 'Debe ingresar su dirección' : null
-            ), */
+              validator: (value) => value == null || value.isEmpty? 'Debe ingresar su dirección' : null
+            ), 
             Container(
               margin: EdgeInsets.symmetric(
                 horizontal: 40.0
@@ -235,7 +235,9 @@ class _RegisterViewState extends State<RegisterView> {
         birthday: birthday['birthday']!,
         createdDate: DateTime.now(),
         lastUpdateDate: DateTime.now(),
-        termsChecked: true
+        termsChecked: true,
+        location: controllers['location']!.text,
+        district: selectedDistrict['district']
       );
 
       final loading = AwesomeDialog(
